More pythonic and added sound controls
Made some of my code more "pythonic" as per the request of some people I shared this with. Will continue to improve this, though. There's still more to go.

Also added a sound module. It will now monitor your sound (requires alsamixer and pyalsaaudio library), and you can control it from the app.
Will need to add list of controls to some sort of help screen, but for now, I will update the readme with them.